fix(test): clear the strict-build red in the rope oracle
Codex's ROCm build with EMBER_STRICT=ON was red on test_qwen_rope_graph_oracle
and it correctly refused to commit on top. Two causes, one of them a real
defect.

Real: three -Wsign-conversion sites. `(static_cast<size_t>(token) * kHeads +
head) * kHeadDim` converts `head` from int into size_t arithmetic. Now cast
explicitly.

Not real, and the target should never have been strict: the remaining
diagnostic is -Wshadow inside vendored ggml-backend.h:355, where a function
name hides ggml_backend_graph_copy's constructor. Calling ember_strict() on a
target that includes vendored engine/ headers is exactly what the
EMBER_STRICT_TARGETS comment says not to do, and it is why
test_qwen4exp_frontier -- which includes the same header -- is not strict.
test_qwen4exp_shards stays strict only because it does not include that header.
Dropped, with the reason recorded at the call site so it is not re-added.

Verified by compiling the translation unit with the exact EMBER_STRICT_FLAGS
set: the three sign-conversion diagnostics are gone and the vendored -Wshadow
is the only one left, which dropping ember_strict excludes.

+++ test/test_qwen_rope_graph_oracle.cpp
@@ -196,7 +196,7 @@ std::vector<float> reference_rope(const ember_qwen_yarn_config & cfg,
         for (int head = 0; head < kHeads; ++head) {
             float * slot =
                 out.data() +
-                (static_cast<size_t>(token) * kHeads + head) * kHeadDim;
+                (static_cast<size_t>(token) * kHeads + static_cast<size_t>(head)) * kHeadDim;
             ember_qwen_yarn_apply(slot, kHeadDim, &cfg, position);
         }
     }
@@ -221,7 +221,7 @@ std::vector<float> exact_rope(const ember_qwen_yarn_config & cfg,
         for (int head = 0; head < kHeads; ++head) {
             float * slot =
                 out.data() +
-                (static_cast<size_t>(token) * kHeads + head) * kHeadDim;
+                (static_cast<size_t>(token) * kHeads + static_cast<size_t>(head)) * kHeadDim;
             for (int k = 0; k < EMBER_QWEN_ROPE_FREQ_COUNT; ++k) {
                 const double angle =
                     std::fmod(pos * exact_inv_freq(cfg, k), 2.0 * 3.14159265358979323846);
@@ -308,7 +308,7 @@ void run_case(bool enable_yarn, int32_t max_context, Path path,
     for (int token = 0; token < kTokens; ++token) {
         for (int head = 0; head < kHeads; ++head) {
             const size_t base =
-                (static_cast<size_t>(token) * kHeads + head) * kHeadDim;
+                (static_cast<size_t>(token) * kHeads + static_cast<size_t>(head)) * kHeadDim;
             for (int i = EMBER_QWEN_ROPE_DIM; i < kHeadDim; ++i) {
                 tail_delta = std::fmax(
                     tail_delta,
